moved vertex_cover functions to a separate file vertex_cover.py
because of circular import issue with bipartite module when placed in covering.py
