+++ networkx/algorithms/covering.py
@@ -4,12 +4,6 @@ from functools import partial
 from itertools import chain
 
 import networkx as nx
-from networkx.algorithms.vertex_covering.lp_decomposition import find_lp_decomposition
-from networkx.algorithms.vertex_covering.vertex_cover import (
-    max_degree_branching,
-    vc_above_lp_branching,
-    vertex_cover_preprocessing,
-)
 from networkx.utils import arbitrary_element, not_implemented_for
 
 __all__ = ["min_edge_cover", "is_edge_cover"]
@@ -146,68 +140,3 @@ def is_edge_cover(G, cover):
     the graph is incident to at least one edge of the set.
     """
     return set(G) <= set(chain.from_iterable(cover))
-
-
-@not_implemented_for("directed")
-@nx._dispatchable
-def vertex_cover(G, k):
-    # find lp-opt value
-    # compare 1.4656^k and 2.618^(k - lpOpt)
-
-    vc = set()
-    G, k, vc, is_k_vc_possible = vertex_cover_preprocessing(G, k, vc)
-
-    if not is_k_vc_possible:
-        return False, set()
-
-    lp_opt_value, *_ = find_lp_decomposition(G, k)
-
-    if lp_opt_value > k:
-        return False, set()
-
-    vc_above_lp_opt_algo_check = 2.618 ** (k - lp_opt_value)
-    max_deg_algo_check = 1.4656**k
-
-    if vc_above_lp_opt_algo_check < max_deg_algo_check:
-        return vc_above_lp_branching(G, k)
-    else:
-        return max_degree_branching(G, k)
-
-
-@not_implemented_for("directed")
-@nx._dispatchable
-def is_vertex_cover(G, cover):
-    """
-    Decides whether a set of vertices is a valid vertex cover for the graph.
-
-    Given a set of vertices, whether it is a vertex cover can be decided if
-    we check that for all edges, atleast one of the end points is present
-    in the set.
-
-    Parameters
-    ----------
-    G : NetworkX graph
-        An undirected bipartite graph.
-
-    cover : set
-        Set of vertices.
-
-    Returns
-    -------
-    bool
-        Whether the set of vertices is a valid vertex cover of the graph.
-
-    Examples
-    --------
-    >>> G = nx.Graph([(0, 1), (0, 2), (0, 3), (1, 2), (1, 3)])
-    >>> cover = {0, 1}
-    >>> nx.is_vertex_cover(G, cover)
-    True
-
-    Notes
-    -----
-    A vertex cover of a graph is a set of vertices such that every edge is
-    incident on atleast one vertex from the set
-
-    """
-    return all(u in cover or v in cover for (u, v) in G.edges)
