+++ networkx/algorithms/vertex_covering/__init__.py
@@ -1,4 +1,4 @@
-from vertex_cover import *
-from crown_decomposition import *
-from lp_decomposition import *
-from vertex_cover_preprocessing import *
+from networkx.algorithms.vertex_covering.vertex_cover import *
+from networkx.algorithms.vertex_covering.crown_decomposition import *
+from networkx.algorithms.vertex_covering.lp_decomposition import *
+from networkx.algorithms.vertex_covering.vertex_cover_preprocessing import *
